fix validation when viewing yaml editor
