+++ datahub-web-react/src/app/ingestV2/source/multiStepBuilder/steps/step2ConnectionDetails/sections/recipeSection/RecipeSection.tsx
@@ -1,4 +1,4 @@
-import { message } from 'antd';
+import { Form, message } from 'antd';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import YAML from 'yamljs';
 
@@ -28,16 +28,31 @@ export function RecipeSection({ state, displayRecipe, sourceConfigs, setStagedRe
         setIsRecipeValid?.(!hasForm || isEditing || !!state.isConnectionDetailsValid);
     }, [hasForm, isEditing, setIsRecipeValid, state.isConnectionDetailsValid]);
 
+    const [form] = Form.useForm();
+    const runFormValidation = useCallback(() => {
+        form.validateFields()
+            .then(() => {
+                setIsRecipeValid?.(true);
+            })
+            .catch((error) => {
+                // FYI: `error` could be triggered with empty list of `errorFields` when form is valid
+                const hasErrors = (error.errorFields?.length ?? 0) > 0;
+                setIsRecipeValid?.(!hasErrors);
+            });
+    }, [form, setIsRecipeValid]);
+
     const onTabClick = useCallback(
         (activeKey) => {
             if (activeKey !== 'form') {
                 setSelectedTabKey(activeKey);
+                setIsRecipeValid?.(true); // no field validation when in yaml editor
                 return;
             }
 
             // Validate yaml content when switching from yaml tab to form
             try {
                 YAML.parse(displayRecipe);
+                setTimeout(runFormValidation, 0); // let form remount and then run validation
                 setSelectedTabKey(activeKey);
             } catch (e) {
                 message.destroy();
@@ -47,7 +62,7 @@ export function RecipeSection({ state, displayRecipe, sourceConfigs, setStagedRe
                 message.warn(`Found invalid YAML. ${messageText} in order to switch views.`);
             }
         },
-        [displayRecipe],
+        [displayRecipe, setIsRecipeValid, runFormValidation],
     );
 
     const tabs: Tab[] = useMemo(
@@ -58,10 +73,11 @@ export function RecipeSection({ state, displayRecipe, sourceConfigs, setStagedRe
                 component: (
                     <RecipeForm
                         state={state}
+                        form={form}
+                        runFormValidation={runFormValidation}
                         displayRecipe={displayRecipe}
                         sourceConfigs={sourceConfigs}
                         setStagedRecipe={setStagedRecipe}
-                        setIsRecipeValid={setIsRecipeValid}
                     />
                 ),
             },
@@ -71,7 +87,7 @@ export function RecipeSection({ state, displayRecipe, sourceConfigs, setStagedRe
                 component: <YamlEditor value={displayRecipe} onChange={setStagedRecipe} />,
             },
         ],
-        [displayRecipe, state, sourceConfigs, setStagedRecipe, setIsRecipeValid],
+        [displayRecipe, state, sourceConfigs, setStagedRecipe, form, runFormValidation],
     );
 
     if (hasForm) {
